requirements.txt
# PyTorch (install first, CUDA 11.8 version)
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0
torchvision>=0.15.0

# Core ML
insightface==0.7.3
onnxruntime-gpu==1.16.0

# MediaPipe for landmarks
mediapipe==0.10.0

# Video/Image processing
opencv-python==4.8.0.76
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0

# Audio handling
ffmpeg-python==0.2.0

# CLI
typer[all]==0.9.0

# Face enhancement
gfpgan==1.3.8
basicsr>=1.4.2
facexlib>=0.3.0
realesrgan>=0.3.0

# Config and utils
pydantic>=2.0.0
tqdm>=4.66.0
typing_extensions>=4.0.0

